Handle spaces in editor command.
Sometimes you want to pass additional args to your editor in $EDITOR
(like when launching mvim with "-f" option to keep running in
foreground). Using it this way would produce a subprocess call like
this:

    subprocess.call(['mvim -f', 'my_filename'])

and python would fail to find the command "mvim -f" because shell=False.

This commit changes it to produce:

    subprocess.call(['mvim', '-f', 'my_filename'])